refactor(bridge): 패키지 구조 변경
새로 controller, core, view 패키지 생성 후 분류
  - controller -> Controller 클래스
  - core -> Bridge, BridgeGame, GameCommand, History 클래스
  - view -> InputView, InputViewRetryProxy, OutputView 클래스
테스트 코드도 동일한 규칙으로 변경
