+++ src/main/java/bridge/core/Bridge.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.core;
 
 import java.util.List;
