+++ src/main/java/bridge/core/BridgeGame.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.core;
 
 import java.util.ArrayList;
 import java.util.List;
