+++ src/main/java/bridge/core/History.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.core;
 
 import java.util.Objects;
